refactor: 일기 수정 test 분리
