+++ apps/api/src/diary/spec/diary.service.spec.ts
@@ -61,103 +61,6 @@ describe('DiaryService', () => {
     uploadFileService = module.get(UploadFileService);
   });
 
-  describe('updateDiary', () => {
-    it('일기장 업데이트 성공', async () => {
-      const mockDiary: DiaryEntity = {
-        id: 1,
-        createdAt: new Date(),
-        updatedAt: new Date(),
-        status: true,
-        userId: 1,
-        title: 'title',
-        content: 'content',
-        isOpen: true,
-      };
-      diaryRepository.findByUnique.mockResolvedValue(mockDiary);
-
-      const result = await diaryService.updateDiary({
-        userId: 1,
-        bodyDto: {
-          title: 'new title',
-          content: 'new content',
-          isOpen: false,
-        },
-        diaryId: 1,
-      });
-
-      expect(result).toBeUndefined();
-      expect(diaryRepository.update).toHaveBeenCalledTimes(1);
-    });
-
-    it('없는 일기장일 경우 NotFoundDiaryException 을 호출한다', async () => {
-      diaryRepository.findByUnique.mockResolvedValue(null);
-
-      await expect(async () => {
-        await diaryService.updateDiary({
-          userId: 1,
-          bodyDto: {
-            title: 'new title',
-            content: 'new content',
-            isOpen: false,
-          },
-          diaryId: 2,
-        });
-      }).rejects.toThrow(new NotFoundDiaryException());
-    });
-
-    it('일기장 status 가 false 인 경우 NotFoundDiaryException 을 호출한다', async () => {
-      const mockDiary: DiaryEntity = {
-        id: 1,
-        createdAt: new Date(),
-        updatedAt: new Date(),
-        status: false,
-        userId: 1,
-        title: 'title',
-        content: 'content',
-        isOpen: true,
-      };
-      diaryRepository.findByUnique.mockResolvedValue(mockDiary);
-
-      await expect(async () => {
-        await diaryService.updateDiary({
-          userId: 1,
-          bodyDto: {
-            title: 'new title',
-            content: 'new content',
-            isOpen: false,
-          },
-          diaryId: 2,
-        });
-      }).rejects.toThrow(new NotFoundDiaryException());
-    });
-
-    it('유저의 일기장이 아닐 경우 NotUserDiaryException 을 호출한다', async () => {
-      const mockDiary: DiaryEntity = {
-        id: 1,
-        createdAt: new Date(),
-        updatedAt: new Date(),
-        status: true,
-        userId: 2,
-        title: 'title',
-        content: 'content',
-        isOpen: true,
-      };
-      diaryRepository.findByUnique.mockResolvedValue(mockDiary);
-
-      await expect(async () => {
-        await diaryService.updateDiary({
-          userId: 1,
-          bodyDto: {
-            title: 'new title',
-            content: 'new content',
-            isOpen: false,
-          },
-          diaryId: 2,
-        });
-      }).rejects.toThrow(new NotUserDiaryException());
-    });
-  });
-
   describe('createDiaryImage', () => {
     it('일기장 이미지 생성 성공', async () => {
       const mockDiary: DiaryEntity = {
